Test Version Wazo check-ssl-certificate.sh

+++ check-ssl-certificate/check-ssl-certificate.sh
@@ -1,5 +1,17 @@
 #!/bin/bash
 
+WAZO_VERSION_FILE=("/usr/share/wazo/WAZO-VERSION")
+MINIMUM_REQUIRED_VERSION="24.01"
+CURRENT_WAZO_VERSION=$(cat "$WAZO_VERSION_FILE")
+
+function is_valid_version() {
+    if [[ $1 =~ ^[0-9]{2}\.[0-9]{2}$ ]]; then
+        return 0
+    else
+        return 1
+    fi
+}
+
 check_expire_date() {
 
 certificate_end_date=$(echo | openssl s_client -connect localhost:443 2> /dev/null | openssl x509 -noout -enddate | cut -d '=' -f2)
@@ -43,7 +55,28 @@ if [ $1 = 0 ]; then
             wazo-service stop all
             echo -e "XXX\n100\nAll Wazo Services stopping... Done.\nXXX"
             sleep 0.7
+    
+    if [ ! -f "$WAZO_VERSION_FILE" ]; then
+        echo "Le fichier $WAZO_VERSION_FILE n'existe pas. Veuillez entrer la version manuellement:"
+        read -r MANUAL_WAZO_VERSION
+    
+        while ! is_valid_version "$MANUAL_WAZO_VERSION"; do
+            echo "La version saisie '$MANUAL_WAZO_VERSION' ne semble pas être au bon format (xx.xx). Merci de recommencer."
+            read -r MANUAL_WAZO_VERSION
+        done
+    else
+        if ("$CURRENT_WAZO_VERSION -le $MINIMUM_REQUIRED_VERSION" || "$MANUAL_WAZO_VERSION -le $MINIMUM_REQUIRED_VERSION")
+            echo -e "XXX\n50\nBackup and remove certificates... Wait...\nXXX"
+            cp /usr/share/xivo-certs/server.{key,crt} /var/backups
+            rm /usr/share/xivo-certs/server.{key,crt}
+            echo -e "XXX\n100\nBackup and remove certificates... Done.\nXXX"
+            sleep 0.7
 
+            echo -e "XXX\n50\nRegenerate self-signed certificate... Wait...\nXXX"
+            dpkg-reconfigure xivo-certs /dev/null 2>&1
+            echo -e "XXX\n100\nRegenerate self-signed certificate... Done.\nXXX"
+            sleep 0.7
+        else
             echo -e "XXX\n50\nBackup and remove certificates... Wait...\nXXX"
             cp /usr/share/wazo-certs/server.{key,crt} /var/backups
             rm /usr/share/wazo-certs/server.{key,crt}
@@ -54,7 +87,8 @@ if [ $1 = 0 ]; then
             dpkg-reconfigure wazo-certs /dev/null 2>&1
             echo -e "XXX\n100\nRegenerate self-signed certificate... Done.\nXXX"
             sleep 0.7
-
+        fi
+    fi
             echo -e "XXX\n50\nRestart Wazo Services... Wait...\nXXX"
             xivo-update-config /dev/null 2>&1
             wazo-service start all
